fix rounding error caused by float division introduced commit b2c27b87

+++ src/netem.c
@@ -184,7 +184,7 @@ int netem_get_params(char *iface, struct netem_params *params)
 		fprintf(stderr, "couldn't get loss for iface: %s\n", iface);
 		goto cleanup_qdisc;
 	}
-	params->loss = (double)loss / ((double)UINT_MAX / 100.0);
+	params->loss = (int)(loss / (UINT_MAX / 100));
 
 	rtnl_qdisc_put(found_qdisc);
 	rtnl_qdisc_put(filter_qdisc);
